Change recursive instance resolving for an iterative version
During the model building, each time a cross-ref towards an other
element is found, a 'CrossRef' element is created and added in the
model graph. Later, the whole graph is scanned and each time a
'CrossRef' is found, the reference is resolved and the 'CrossRef' is
replaced by the resolved element.

This new version uses the same cross-ref resolution method, but keeps
track of the unresolved cross-ref, and resolves them later. The
resolution thus uses a simple for-loop on the cross-ref list instead
of the recursive algorithm.

Performance measurement from the 'tests/perf' folder shows a small
difference between the recursive and the iterative version. Tests have
been made with Python 3.6.2 with an i7 processor. The iterative
version is up to 1s faster than the recursive version on big models.