Rework textX CLI extensibility to be simpler and more flexible
Now you should register a Python function that accepts the top-level `textx`
command group and use that to register additional commands using `click`
decorators. In this way we use the same mechanism for both a plain command and
command group registration. In addition it is now possible to provide additional
parameters to `click` command parser which will be useful in advanced textX
commands.